Azure_AI/agent_tools_with_mcp/requirements.txt
uvicorn
azure-ai-projects
azure-identity
python-dotenv

# pip install -r requirements.txt --pre azure-ai-projects mcp